refactor(workspace): split Workspace.tsx into per-component modules
Move-only extraction of sub-components, shared types/constants/helpers into
pages/Workspace/ so the main Workspace() component is easier to navigate.
