+++ frontend/src/pages/Workspace/chat/ChatMessageBody.tsx
@@ -1,0 +1,28 @@
+// Renders a single chat message body (Markdown for assistant, plain text otherwise).
+// Parent: ChatPanel.
+
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+
+import type { WorkspaceMessage } from '../types';
+
+export function ChatMessageBody({ message }: { message: WorkspaceMessage }) {
+  const isMarkdown = message.role === 'assistant' && message.kind !== 'tool_log';
+  if (!isMarkdown) {
+    return <>{message.content}</>;
+  }
+  return (
+    <div className="workspace-chat-markdown">
+      <ReactMarkdown
+        remarkPlugins={[remarkGfm]}
+        components={{
+          a: ({ node, ...props }) => (
+            <a {...props} target="_blank" rel="noopener noreferrer" />
+          ),
+        }}
+      >
+        {message.content}
+      </ReactMarkdown>
+    </div>
+  );
+}
